correct compilation with clang

+++ Source/source/base/MathOperators.cpp
@@ -33,7 +33,8 @@ namespace optim
     _BitScanReverse(&result, static_cast<u32>(x));
     return(result);
 #else
-    return(32 - __builtin_clz(x));
+    return(32u - static_cast<u32>(__builtin_clz(x)));
 #endif
 }
 }
+
